+++ be1-go/internal/handler/messagedata/coin/coin_test.go
